+++ src/Builder/Predicate/Where.php
@@ -1,0 +1,110 @@
+<?php
+
+namespace Bank\Builder\Predicate;
+
+use Bank\Builder\PredicateBuilder;
+use Bank\Query\Predicate\Where as WhereQuery;
+
+/**
+ * Class Where
+ * @package Bank\Builder\Predicate
+ */
+class Where extends PredicateBuilder
+{
+
+    const WHERE_CLAUSE = "WHERE";
+
+    /**
+     * @param WhereQuery $where
+     * @return string
+     */
+    public function build($where): string
+    {
+        $conditions = $where->getConditions();
+
+        if (!$conditions) {
+            return "";
+        }
+
+        $query = array_reduce($conditions, function ($query, $condition) {
+
+            $value = $condition["value"] ?? null;
+            $join = $condition["join"] ?? null;
+
+            if ($value instanceof WhereQuery) {
+                return $this->buildNestWhere($query, $value, $join);
+            }
+
+            return $this->buildSimpleWhere($query, $condition);
+        }, []);
+
+        $where = implode("", $query);
+
+        return $where;
+    }
+
+    /**
+     * @param array $query
+     * @param array $condition
+     * @return array
+     */
+    protected function buildSimpleWhere(array $query, array $condition): array
+    {
+        $conditionVal = $condition["value"] ?? null;
+        $col = $condition["col"] ?? null;
+        $operator = $condition["operator"] ?? null;
+        $join = $condition["join"] ?? null;
+
+        $value = $this->castWhereValue($conditionVal);
+
+        list($table, $column) = $this->divideFirstParam($col);
+        if ($table) {
+            $col = $table . "." . $column;
+        }
+
+        $where = ($value) ? "{$col} {$operator} {$value}" : "{$col} {$operator}";
+        if ($query) {
+            $where = " {$join} {$where}";
+        }
+
+        $query[] = $where;
+
+        return $query;
+    }
+
+    /**
+     * @param array $query
+     * @param Where $where
+     * @param string $join
+     * @return array
+     */
+    protected function buildNestWhere(array $query, WhereQuery $where, string $join): array
+    {
+        $nest = $this->enclosedInBracket($this->build($where));
+        if ($query) {
+            $nest = " {$join} {$nest}";
+        }
+        $query[] = $nest;
+        return $query;
+    }
+
+    /**
+     * @param $conditionVal
+     * @return string
+     */
+    protected function castWhereValue($conditionVal): string
+    {
+        $value = "";
+        if (is_array($conditionVal)) {
+            $value = array_map(function ($item) {
+                return $this->quote($item);
+            }, $conditionVal);
+
+            $value = $this->enclosedInBracket(implode(" , ", $value));
+        } else if (!empty($conditionVal)) {
+            $value = $this->quote($conditionVal);
+        }
+
+        return $value;
+    }
+}
