+++ src/Builder/PredicateBuilderInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Bank\Builder;
+
+/**
+ * Interface PredicateBuilderInterface
+ * @package Bank\Builder
+ */
+interface PredicateBuilderInterface
+{
+    /**
+     * @param $query
+     * @return string
+     */
+    public function build($query): string;
+
+}
